All tests on one file
