+++ test/LifToken.js
@@ -1,3 +1,5 @@
+
+var abi = require('ethereumjs-abi');
 var protobuf = require("protobufjs");
 
 String.prototype.hexEncode = function(){
@@ -53,46 +55,165 @@ contract('LifToken', function(accounts) {
     done();
   });
 
-  function checkValues (_totalSupply, _feesBalance, tokenPrice, tokenFee, _accounts, done) {
-    var accountPromises = [];
-    accountPromises.push( web3.eth.getBalance(token.contract.address) );
-    accountPromises.push( token.totalSupply() );
-    accountPromises.push( token.feesBalance() );
-    accountPromises.push( token.tokenPrice() );
-    accountPromises.push( token.tokenFee() );
+  function chekValues (_totalSupply, _feesBalance, tokenPrice, tokenFee, _accounts) {
+    return new Promise(function(resolve, reject) {
 
-    for (var i = 0; i < _accounts.length; i++) {
-      accountPromises.push( token.balanceOf(accounts[i]) );
-    }
+      var accountPromises = [];
+      accountPromises.push( web3.eth.getBalance(token.contract.address) );
+      accountPromises.push( token.totalSupply() );
+      accountPromises.push( token.feesBalance() );
+      accountPromises.push( token.tokenPrice() );
+      accountPromises.push( token.tokenFee() );
 
-    Promise.all(accountPromises).then(values => {
+      for (var i = 0; i < _accounts.length; i++) {
+        accountPromises.push( token.balanceOf(accounts[i]) );
+      }
 
-      if (DEBUG_MODE) {
-        console.log('Contract Balance:', web3.fromWei(parseInt(values[0]), 'ether'), 'Ether;', parseInt(values[0], 'wei'), 'Wei');
-        console.log('Total Supply:', parseBalance(values[1]));
-        console.log('Fees Balance:', web3.fromWei(parseInt(values[2]), 'ether'), 'Ether;', parseInt(values[2], 'wei'), 'Wei');
-        console.log('Token Price:', parseInt(values[3]));
-        console.log('Token Fee:', parseInt(values[4]));
+      Promise.all(accountPromises).then(values => {
 
-        for (var z = 5; z < values.length; z++) {
-          console.log('Account['+(z-5)+']', accounts[z-5], ", Balance:", parseBalance(values[z]));
+        if (DEBUG_MODE) {
+          console.log('Contract Balance:', web3.fromWei(parseInt(values[0]), 'ether'), 'Ether;', parseInt(values[0], 'wei'), 'Wei');
+          console.log('Total Supply:', parseBalance(values[1]));
+          console.log('Fees Balance:', web3.fromWei(parseInt(values[2]), 'ether'), 'Ether;', parseInt(values[2], 'wei'), 'Wei');
+          console.log('Token Price:', parseInt(values[3]));
+          console.log('Token Fee:', parseInt(values[4]));
+
+          for (var z = 5; z < values.length; z++)
+            console.log('Account['+(z-5)+']', accounts[z-5], ", Balance:", parseBalance(values[z]));
+
         }
-      }
 
-      assert.equal(parseBalance(values[1]), _totalSupply);
-      assert.equal(web3.fromWei(parseInt(values[2])), _feesBalance);
-      assert.equal(parseInt(values[3]), tokenPrice);
-      assert.equal(parseInt(values[4]), tokenFee);
+        assert.equal(parseBalance(values[1]), _totalSupply);
+        assert.equal(web3.fromWei(parseInt(values[2])), _feesBalance);
+        assert.equal(parseInt(values[3]), tokenPrice);
+        assert.equal(parseInt(values[4]), tokenFee);
 
-      for (var x = 5; x < values.length; x++) {
-        assert.equal(parseBalance(values[x]), _accounts[x-5]);
-      }
+        for (var x = 5; x < values.length; x++)
+          assert.equal(parseBalance(values[x]), _accounts[x-5]);
 
-      done();
-    }).catch(err => {
-      done(err);
+        resolve();
+      }).catch(err => {
+        reject(err);
+      });
     });
   }
+
+
+  function getActions() {
+    return new Promise(function(resolve, reject) {
+
+      token.DAOActionsLength().then(actionsLenght => {
+
+        var actionPromises = [];
+
+        for (var z = 1; z < actionsLenght; z++)
+          actionPromises.push( token.DAOActions.call(z) );
+
+        Promise.all(actionPromises).then(actions => {
+
+          if (DEBUG_MODE){
+            console.log('Total Actions:', parseInt(actionsLenght)-1);
+            for (var z = 0; z < actions.length; z++)
+              console.log('Signature:', actions[z][2], '; Address:', actions[z][0], '; % Votes:', parseInt(actions[z][1]));
+          }
+          resolve(actions);
+        }).catch(err => {
+          reject(err);
+        });
+      });
+
+    });
+  }
+
+  function getProposals() {
+    return new Promise(function(resolve, reject) {
+
+      token.ProposalsLenght().then(proposalsLenght => {
+        var actionPromises = [];
+
+        for (var z = 1; z < proposalsLenght; z++)
+          actionPromises.push( token.proposals.call(z) );
+
+        Promise.all(actionPromises).then(proposals => {
+          if (DEBUG_MODE){
+            console.log('Total Proposals:', parseInt(proposalsLenght)-1);
+            for (var z = 0; z < proposals.length; z++)
+              console.log('['+parseInt(proposals[z][1])+'] To: '+proposals[z][0]+', Value: '+web3.fromWei(proposals[z][2],'ether')+', Desc: '+proposals[z][3]+', Status: '+parseInt(proposals[z][4]));
+          }
+          resolve(proposals);
+        }).catch(err => {
+          reject(err);
+        });
+      });
+
+    });
+  }
+
+  function getProposal(id) {
+    return new Promise(function(resolve, reject) {
+
+      token.proposals.call(id).then(proposal => {
+        var votesPromises = [];
+
+        var parsedProposal = {
+          target: proposal[0],
+          id: parseInt(proposal[1]),
+          value: parseInt(proposal[2]),
+          description: proposal[3],
+          status: parseInt(proposal[4]),
+          creationBlock: parseInt(proposal[5]),
+          maxBlock: parseInt(proposal[6]),
+          executionBlock: parseInt(proposal[7]),
+          approvalVotes: parseInt(proposal[8]),
+          actionData: proposal[9],
+          totalVotes: parseInt(proposal[10]),
+          positiveVotes: 0,
+          votesNeeded: 0,
+          votes: []
+        };
+
+        for (var z = 1; z < parsedProposal.totalVotes+1; z++)
+          votesPromises.push(token.getProposalVote(parsedProposal.id, z));
+
+        token.totalSupply().then(totalSupply => {
+          parsedProposal.votesNeeded = (parsedProposal.approvalVotes / 100) * parseBalance(parseInt(totalSupply));
+
+          Promise.all(votesPromises).then(votesDone => {
+            for (var i = 0; i < votesDone.length; i++){
+              parsedProposal.votes.push({
+                address: votesDone[i].voter,
+                amount: parseBalance(parseInt(votesDone[i].balance)),
+                vote: votesDone[i].vote
+              });
+              if (votesDone[i][2])
+                parsedProposal.positiveVotes += parseBalance(parseInt(votesDone[i][1]));
+            }
+            console.log('['+parsedProposal.id+'] To: '+parsedProposal.target+', Value: '+parsedProposal.value+', Desc: '+parsedProposal.description+', Status: '+parsedProposal.status, ', Votes: ',parsedProposal.positiveVotes,'/',parsedProposal.votesNeeded);
+            resolve();
+          }).catch(err => {
+            reject(err);
+          });
+        });
+      });
+
+    });
+  }
+
+  function waitBlocks(toWait){
+    return new Promise(function(resolve, reject) {
+      var toWait =+ web3.eth.blockNumber;
+      var wait = setInterval( function() {
+          if (web3.eth.blockNumber >= toWait) {
+              clearInterval(wait);
+              resolve();
+          }
+      }, 1000 );
+    });
+  }
+
+////////////////////////////////////////////////////////
+//                    Lif Token Tests                 //
+////////////////////////////////////////////////////////
 
   it("should return the correct totalSupply after construction using createTokens", function(done) {
     return token.start()
@@ -100,7 +221,9 @@ contract('LifToken', function(accounts) {
         return token.createTokens(accounts[0], { value: web3.toWei(1, 'ether') });
       })
       .then(function() {
-        checkValues(100, 0, 10000000000000000, 100, [100, 0, 0], done);
+        return chekValues(100, 0, 10000000000000000, 100, [100, 0, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -113,7 +236,9 @@ contract('LifToken', function(accounts) {
         if (error.message.search('invalid JUMP') == -1) throw error;
       })
       .then(function() {
-        checkValues(0, 0, 10000000000000000, 100, [0, 0, 0], done);
+        return chekValues(0, 0, 10000000000000000, 100, [0, 0, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -130,7 +255,9 @@ contract('LifToken', function(accounts) {
       })
       .then(function(allowance) {
         assert.equal(parseBalance(allowance), 100);
-        checkValues(100, 0, 10000000000000000, 100, [100, 0, 0], done);
+        return chekValues(100, 0, 10000000000000000, 100, [100, 0, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -143,7 +270,9 @@ contract('LifToken', function(accounts) {
         return token.transfer(accounts[1], formatBalance(33), "");
       })
       .then(function() {
-        checkValues(99.67, 0.000000000000000003, 10000000000000000, 100, [67, 32.67, 0], done);
+        return chekValues(99.67, 0.000000000000000003, 10000000000000000, 100, [67, 32.67, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -159,7 +288,9 @@ contract('LifToken', function(accounts) {
         if (error.message.search('invalid JUMP') == -1) throw error;
       })
       .then(function() {
-        checkValues(100, 0, 10000000000000000, 100, [100, 0, 0], done);
+        return chekValues(100, 0, 10000000000000000, 100, [100, 0, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -175,7 +306,9 @@ contract('LifToken', function(accounts) {
         return token.transferFrom(accounts[0], accounts[2], formatBalance(100), "", {from: accounts[1]});
       })
       .then(function() {
-        checkValues(99, 0.000000000000000001, 10000000000000000, 100, [0, 0, 99], done);
+        return chekValues(99, 0.000000000000000001, 10000000000000000, 100, [0, 0, 99]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -194,7 +327,9 @@ contract('LifToken', function(accounts) {
         if (error.message.search('invalid JUMP') == -1) throw error;
       })
       .then(function() {
-        checkValues(100, 0, 10000000000000000, 100, [100, 0, 0], done);
+        return chekValues(100, 0, 10000000000000000, 100, [100, 0, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -207,7 +342,9 @@ contract('LifToken', function(accounts) {
         if (error.message.search('invalid JUMP') == -1) throw error;
       })
       .then(function() {
-        checkValues(0, 0, 10000000000000000, 100, [0, 0, 0], done);
+        return chekValues(0, 0, 10000000000000000, 100, [0, 0, 0]);
+      }).then(function(){
+        done();
       });
   });
 
@@ -227,8 +364,10 @@ contract('LifToken', function(accounts) {
           assert.equal(error, null);
           var decodedObj = JSON.parse(log[0].args.data.toString().hexDecode());
           assert.equal("AwesomeString", decodedObj.awesomeField);
-          checkValues(99, 0.000000000000000001, 10000000000000000, 100, [0, 99, 0], done);
+          return chekValues(99, 0.000000000000000001, 10000000000000000, 100, [0, 99, 0]);
         });
+      }).then(function(){
+        done();
       });
   });
 
@@ -254,8 +393,532 @@ contract('LifToken', function(accounts) {
           assert.equal(error, null);
           var decodedBuffer = new Buffer(log[0].args.data.toString().hexDecode());
           assert.equal("AwesomeString", AwesomeMessage.decode(decodedBuffer).awesomeField);
-          checkValues(10, 0, 10000000000000000, 100, [10, 0, 0], done);
+          return chekValues(10, 0, 10000000000000000, 100, [10, 0, 0]);
         });
+      }).then(function(){
+        done();
+      });
+  });
+
+////////////////////////////////////////////////////////
+//                    Lif DAO Tests                   //
+////////////////////////////////////////////////////////
+
+  it("Should add the min votes needed for native contract actions", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setPrice(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    console.log('Action setPrice(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 85, signature)
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setFee(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 86, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setBaseProposalFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setBaseProposalFee(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 87, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setProposalAmountFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setProposalAmountFee(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 88, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setMaxSupply(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setMaxSupply(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 89, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setSpecialActionMinVotes(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setSpecialActionMinVotes(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 90, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setMigrationMinVotes(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setMigrationMinVotes(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 91, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setProposalBlocksWait(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action setProposalBlocksWait(uint256) signature', signature);
+        return token.buildMinVotes(token.contract.address, 92, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "claimFees(address)", web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action claimFees(address) signature', signature);
+        return token.buildMinVotes(token.contract.address, 50, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "addDAOAction(address,uint,bytes4)", web3.toHex(0), web3.toHex(0), web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action addDAOAction(address,uint,bytes4) signature', signature);
+        return token.buildMinVotes(token.contract.address, 61, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "removeDAOAction(address,bytes4)", web3.toHex(0), web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action removeDAOAction(address,bytes4) signature', signature);
+        return token.buildMinVotes(token.contract.address, 62, signature);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "changeDaoAction(address,uint,bytes4)", web3.toHex(0), web3.toHex(0), web3.toHex(0)).toString('hex').substring(0,10);
+        console.log('Action changeDaoAction(address,uint,bytes4) signature', signature);
+        return token.buildMinVotes(token.contract.address, 63, signature);
+      })
+      .then(function() {
+        return getActions();
+      })
+      .then(function(actions){
+        assert.equal(actions.length, 12);
+        done();
+      });
+  });
+
+  it("Should add a setFee proposal", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    console.log('Action setFee(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], { value: web3.toWei(1, 'ether') });
+      })
+      .then(function() {
+        return chekValues(100, 0, 10000000000000000, 100, [100, 0, 0], null);
+      })
+      .then(function() {
+        var data = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(50)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set fee to 50', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        return getActions();
+      })
+      .then(function(actions){
+        assert.equal(actions.length, 1);
+        return chekValues(100, 100, 10000000000000000, 100, [0, 0, 0], null);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should fail adding setFee proposal due to low value sent", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    console.log('Action setFee(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], { value: web3.toWei(0.1, 'ether') });
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        var data = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(50)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set fee to 50', 0, signature, data, {from: accounts[0]});
+      })
+      .catch(function(error) {
+        if (error.message.search('invalid JUMP') == -1) throw error;
+        done();
+      });
+  });
+
+  it("Shouldnt add the proposal from an address because sender isnt a token holder", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    console.log('Action setFee(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], { value: web3.toWei(1, 'ether') });
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        var data = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(50)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set fee to 50', 0, signature, data, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 0);
+        return getActions();
+      })
+      .then(function(actions){
+        assert.equal(actions.length, 1);
+        done();
+      });
+  });
+
+  it("Should add a setPrice proposal, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setPrice(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    console.log('Action setPrice(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setPrice(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        var data = '0x'+abi.simpleEncode( "setPrice(uint256)", web3.toWei(0.0005, 'ether')).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set price to 0.0005 ether', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function(){
+        return chekValues(610, 100, web3.toWei(0.0005, 'ether'), 100, [10, 500, 0]);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should add a setFee proposal, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    console.log('Action setFee(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var signature = '0x'+abi.simpleEncode( "setFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+        var data = '0x'+abi.simpleEncode( "setFee(uint256)", 50).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set price to 0.0005 ether', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function(){
+        return chekValues(610, 100, 10000000000000000, 50, [10, 500, 0]);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should add a setBaseProposalFee proposal, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setBaseProposalFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    console.log('Action setBaseProposalFee(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var data = '0x'+abi.simpleEncode("setBaseProposalFee(uint256)", web3.toHex(60000000000000000000)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set baseProposalFee to 60 Lif', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.baseProposalFee();
+      })
+      .then(function(baseProposalFee){
+        console.log('New base proposal fee:',parseInt(baseProposalFee));
+        assert.equal(parseInt(baseProposalFee), 60000000000000000000);
+        return chekValues(610, 100, 10000000000000000, 100, [10, 500, 0]);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should add a setProposalAmountFee proposal, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setProposalAmountFee(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    console.log('Action setProposalAmountFee(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var data = '0x'+abi.simpleEncode("setProposalAmountFee(uint256)", web3.toHex(20)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set proposalAmountFee to 2%', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.proposalAmountFee();
+      })
+      .then(function(proposalAmountFee){
+        console.log('New proposal amountFee fee:',parseInt(proposalAmountFee));
+        assert.equal(parseInt(proposalAmountFee), 20);
+        return chekValues(610, 100, 10000000000000000, 100, [10, 500, 0]);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should add a setMaxSupply proposal, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setMaxSupply(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    console.log('Action setMaxSupply(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var data = '0x'+abi.simpleEncode("setMaxSupply(uint256)", web3.toHex(15000000)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set setMaxSupply to 15000000 Lif', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.maxSupply();
+      })
+      .then(function(maxSupply){
+        console.log('New max supply:',parseInt(maxSupply));
+        assert.equal(parseInt(maxSupply), 15000000);
+        return chekValues(610, 100, 10000000000000000, 100, [10, 500, 0]);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should add a setProposalBlocksWait proposal, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode( "setProposalBlocksWait(uint256)", web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    console.log('Action setProposalBlocksWait(uint256) signature', signature);
+    return token.buildMinVotes(token.contract.address, 50, signature)
+      .then(function() {
+        return token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var data = '0x'+abi.simpleEncode("setProposalBlocksWait(uint256)", web3.toHex(666)).toString('hex');
+        return token.newProposal(token.contract.address, 0, 'Set setProposalBlocksWait to 666 blocks', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.proposalBlocksWait();
+      })
+      .then(function(proposalBlocksWait){
+        console.log('New base proposal fee:',parseInt(proposalBlocksWait));
+        assert.equal(parseInt(proposalBlocksWait), 666);
+        return chekValues(610, 100, 10000000000000000, 100, [10, 500, 0]);
+      }).then(function(){
+        done();
+      });
+  });
+
+  it("Should add a proposal to call a function outside the contract, be voted by another user, check it and get executed.", function(done) {
+    var signature = '0x'+abi.simpleEncode("message(bytes32,uint256,string)", web3.toHex(0), web3.toHex(0),web3.toHex(0)).toString('hex').substring(0,10);
+    var proposalID;
+    var test;
+    console.log('Action message(bytes32,uint256,string) signature', signature);
+    return Message.new()
+      .then(function(_message) {
+        message = _message;
+        return token.buildMinVotes(message.contract.address, 50, signature);
+      })
+      .then(function() {
+        token.start();
+      })
+      .then(function() {
+        return token.createTokens(accounts[0], {from: accounts[0], value: web3.toWei(1.1, 'ether') });
+      })
+      .then(function() {
+        return token.createTokens(accounts[1], {from: accounts[1], value: web3.toWei(5, 'ether') });
+      })
+      .then(function() {
+        return chekValues(610, 0, 10000000000000000, 100, [110, 500, 0], null);
+      })
+      .then(function() {
+        var data = '0x'+abi.simpleEncode("showMessage(bytes32,uint256,string)", web3.toHex('Test Bytes32'), web3.toHex(666), 'Test String').toString('hex');
+        return token.newProposal(message.contract.address, 0, 'Call showMessage(bytes32,uint256,string)', 0, signature, data, {from: accounts[0]});
+      })
+      .then(function() {
+        return getProposals();
+      })
+      .then(function(proposals) {
+        assert.equal(proposals.length, 1);
+        proposalID = parseInt(proposals[0][1]);
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return token.vote(proposalID, true, {from: accounts[1]});
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        return token.checkProposal(proposalID);
+      })
+      .then(function() {
+        return getProposal(proposalID);
+      })
+      .then(function() {
+        message.allEvents().get(function(error, log){
+          assert.equal(log[0].event, 'Show');
+          assert.equal(log[0].args.b32, '0x5465737420427974657333320000000000000000000000000000000000000000');
+          assert.equal(parseInt(log[0].args.number), 666);
+          assert.equal(log[0].args.text, 'Test String');
+          return chekValues(610, 100, 10000000000000000, 100, [10, 500, 0]);
+        });
+      }).then(function(){
+        done();
       });
   });
